# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🎄Api Пензенского государственного университета",
        "description": "Api Пензенского государственного университета",
        "version": "Pnzgu"
    },
    "paths": {
...
        "/v2/session/authentication": {
            "post": {
                "tags": [
                    "session"
                ],
                "summary": "Аутентификация пользователей ПГУ",
                "description": "Аутентификация пользователей ПГУ",
                "operationId": "b3afcc28fdaf5178108b150bca4af154",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SessionAuthenticationRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sub": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "517671772"
                                                    },
                                                    "adm": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "pssch": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "0"
                                                    },
                                                    "iat": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733806994"
                                                    },
                                                    "exta": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733893394"
                                                    },
                                                    "extr": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733979794"
                                                    },
                                                    "a_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "eyJhbGciOiJzaGEyNTYiLCJ0eXAiOiJKV1QifQ==.eyJzdWIiOjUxNzY3MTc3MiwiYWRtIjoxLCJwc3NjaCI6MCwiaWF0IjoxNzMzODA2OTk0LCJleHRhIjoxNzFkljhpSlfgJleHRyIjoxNzMzOTc5Nzk0fQ==.8898d5f990bd718a7e3b3afad1be593c26f99b35803797fa15f825b69346a975"
                                                    },
                                                    "r_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "91a04f11291724e9238c2b49549c68d4"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v2/session/refresh": {
            "post": {
                "tags": [
                    "session"
                ],
                "summary": "Обновление a_token пользователей ПГУ",
                "description": "Обновление a_token пользователей ПГУ",
                "operationId": "a76d2b3e27f80c46aa6c832e26dd4198",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SessionRefreshRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "a_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "eyJhbGciOiJzaGEyNTYiLCJ0eXAiOiJKV1QifQ==.eyJzdWIiOjUxNzY3MTc3MiwiYWRtIjoxLCJwc3NjaCI6MCwiaWF0IjoxNzMzODA2OTk0LCJleHRhIjoxNzFkljhpSlfgJleHRyIjoxNzMzOTc5Nzk0fQ==.8898d5f990bd718a7e3b3afad1be593c26f99b35803797fa15f825b69346a975"
                                                    },
                                                    "r_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "91a04f11291724e9238c2b49549c68d4"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v2/session/check": {
            "post": {
                "tags": [
                    "session"
                ],
                "summary": "Проверка a_token пользователей ПГУ",
                "description": "Проверка a_token пользователей ПГУ",
                "operationId": "f179b42c77a421c9676e12d1b1cb9c3a",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SessionCheckRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sub": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "517671772"
                                                    },
                                                    "adm": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "pssch": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "0"
                                                    },
                                                    "iat": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733806994"
                                                    },
                                                    "exta": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733893394"
                                                    },
                                                    "extr": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733979794"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v2/session/change": {
            "patch": {
                "tags": [
                    "session"
                ],
                "summary": "Изменение сессии пользователя",
                "description": "Изменение сессии пользователя",
                "operationId": "587fbe63d66dd4b4a4ab37d4fcb0fcc9",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SessionChangeRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sub": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "517671772"
                                                    },
                                                    "adm": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "pssch": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "0"
                                                    },
                                                    "iat": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733806994"
                                                    },
                                                    "exta": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733893394"
                                                    },
                                                    "extr": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1733979794"
                                                    },
                                                    "a_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "eyJhbGciOiJzaGEyNTYiLCJ0eXAiOiJKV1QifQ==.eyJzdWIiOjUxNzY3MTc3MiwiYWRtIjoxLCJwc3NjaCI6MCwiaWF0IjoxNzMzODA2OTk0LCJleHRhIjoxNzFkljhpSlfgJleHRyIjoxNzMzOTc5Nzk0fQ==.8898d5f990bd718a7e3b3afad1be593c26f99b35803797fa15f825b69346a975"
                                                    },
                                                    "r_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "91a04f11291724e9238c2b49549c68d4"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v2/session/logout": {
            "post": {
                "tags": [
                    "session"
                ],
                "summary": "Выход из аккаунта для пользователей ПГУ",
                "description": "Выход из аккаунта для пользователей ПГУ",
                "operationId": "a74aaec20f4cf78ae4650832ec550a7a",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SessionLogoutRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "a_token": {
                                                        "type": "string",
                                                        "format": "jwt",
                                                        "example": "eyJhbGciOiJzaGEyNTYiLCJ0eXfAiOiJKV1QifQ.eyJzdWIiOjUxNzY3r3MTc3MiwiYWRtIjhoxLCJwc3NjaCI6MCwiaWF0IjoxNzMyNjE3MDE5LCJleHRhIjoxNzMyNzAzNDE5LCJleHRyIjoxNzMyNzg5ODE5fQ.0ff71fc0d98dd36f1df07c0f77062160f26d9a4be1872d407e157fdc85f9f3dd"
                                                    },
                                                    "r_token": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "eyJhbGciOiJzaGEyN"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error500"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error500": {
                "description": "Стандартный формат ошибки",
                "properties": {
                    "status": {
                        "type": "string",
                        "format": "string",
                        "example": "Error"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "Message": {
                                    "type": "string",
                                    "format": "string",
                                    "example": "syntax error, unexpected token \"}\""
                                },
                                "Info": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "requestBodies": {
...
            "SessionAuthenticationRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "login": {
                                    "schema": "required|string",
                                    "type": "string",
                                    "format": "login",
                                    "example": "s517671772"
                                },
                                "password": {
                                    "schema": "required|string",
                                    "type": "string",
                                    "format": "password",
                                    "example": "fdgdhd45"
                                },
                                "platform": {
                                    "schema": "required|int|string",
                                    "type": "string",
                                    "format": "string",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}